remove lies_we_tell_hack.hhi from HSL (#83)
Summary:
Pull Request resolved: https://github.com/facebook/fbshipit/pull/83

Pull Request resolved: https://github.com/hhvm/hsl-experimental/pull/76

Pull Request resolved: https://github.com/hhvm/hsl/pull/98

In HSL and open-source projects that depend on it, most of these "lies" are unneeded, and the remaining places that depend on them can be avoided by changing `\HH\InvariantException` to `InvariantException`. Without namespace, it will always be correctly resolved to wherever HHVM autoimports it from.

"wherever HHVM autoimports it from" is actually changing (it used to be `\InvariantException` but is `\HH\InvariantException` as of a few days ago), which is the reason why this file is problematic. Before the HHVM change, and without the diff, the file was needed -- but after the HHVM change, the file causes Hack errors in HSL and [anything that depends on it](https://travis-ci.org/hhvm/)).

Reviewed By: fredemmott

Differential Revision: D18121499

fbshipit-source-id: abd73c9f72401755ae1250302bb4c515faf23e75