FBCodeAutocargoPlugin: make it possible to import public Cargo.toml changes back to fbcode
Summary:
When a GitHub PR is created the changes to Cargo.toml file has to be mapped somewhere. Since in fbsource there are both public and private Cargo.toml files the tool doesn't know which one to map to.
This chagne will make it so that it will always apply the public Cargo.toml changes to the public Cargo.toml files in fbsource - making the merge clean and obvious for FB employers.

Reviewed By: farnz

Differential Revision: D18780577

fbshipit-source-id: bfab345f4f82dc73ebe4a127c39cc9db53702fb6