fix isSubmoduleDiff regex
Summary:
I noticed in my `git diff` output that there is this "160000" mode string on the "index" line sometimes.

This may be why D19314673 got imported incorrectly (still waiting for my manual import to finish to confirm).

Either way the new regex is a strict superset of the old one so it can't make it worse :)

Reviewed By: zpao

Differential Revision: D19321488

fbshipit-source-id: d9f8f749894ce9453fb0fdc773720be054131c7e