handling edge cases on tag stripping
Summary:
This is a follow-up to D19394534.

All PR imports were getting "[pytorch][PR]" prefixed on them when syncing out to OSS, which wasn't a great look (those tags are still good for internal diffs, though).

This patch changes the logic so that the flag switches between "strip everything" and "strip only the import tags".

...we might also want to consider stripping "[pytorch]" tags in general from the PyTorch repo, because that's a thing which is useful to people in internal diffs but pretty redundant on GitHub.

Reviewed By: zpao

Differential Revision: D19415732

fbshipit-source-id: 804db00cfa455c80052945b104347628a39de022